requirements.txt
absl-py==2.1.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.2.25
exceptiongroup==1.2.2
fastapi==0.103.2
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.33.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.7
importlib-metadata==6.7.0
keras==2.11.0
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
mysqlclient==2.1.0
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-multipart==0.0.8
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.27.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
